I am submitting this movie as a test case for several reasons.  Please delete it if this is not appropriate (and then complain to Bastiaan, who told me to just add it ;).
This is a large movie, and therefore a way to test
performance.  Also, it plays rather badly in Gnash,
and consequently highlights some places for improvement.
Unfortunately it does not include sound.

The original slide presentation was created by Helen
Cook and me; we signed statements releasing the speech
(and filming of the presentation) in to the custody of
The Perl Foundation, which offers material under GPL2 or
the Artistic License (user's choice).  The movie was
exported from Keynote by (I think) Ricardo Signes.
It was created as a lightning talk for YAPC::NA 2006.

Possible concerns would be the closing police car
image, which was found online.  All other images were
taken at the conference for use in this presentation.