Because of John's recent mail concerning FSF copyright/license notes, I added a test which checks for potential problems with the copyright.
It doesn't yet check for odd whitespace after the notice, but it does
check if the copyright exists, if the file starts with a one-line comment
describing the file, if 'You should have received...' begins a new
paragraph, and if the copyright notice includes the current year.

My thought was that perhaps at some moment when there's a minimal chance
of commit conflicts I or someone else could use the test to identify
problem files and correct them.