Big changes here:
Moved this from an article to a book - it seems more appropriate, especially as this document grows with time.

I made a lot of small tagging corrections throughout the files (to be committed shortly) so that gnash.xml validates correctly now.  There are a few more changes on the way, such as using a proper glossary rather than a section titled glossary.  I'll also expand the glossary at that time.

I also still need to do some indexing, and to refresh my memory on how indexes are auto-created with DocBook.