Big changes here:
Added a requirements section; although they aren't well defined at the moment,it's important to pass that along to users (Look! We're flexible!).  Also, moved the dependency information to the front of the installation process.  It's much easier to proceed with the rest of the installation if your target system is already prepared for the Gnash software.