libmedia/Makefile.am,MediaBuffer.h: build and link FFMPEG part of libmedia.
Libmedia is still not widely used, but it does more or less work as it is. The
main obstacle to implementing it in NetStream, where it would improve performance
in a number of ways, is that it requires a tu_file (rather than a LoadingThread)
from NetConnection. That's obviously not hard to hack up for test purposes, but
isn't in the least safe as long as the tu_file in question is still 'owned' by
the loading thread, which other stuff relies on (not to mention NetStreamGst).