* server/sprite_instance.cpp: don't use SIZET_FMT or c_str() in logging 	  calls.
I've replaced SIZET_FMT with %d and dropped c_str() for all logging calls.
This is fine with boost::format. sprite_instance.cpp is one of the heaviest
users of size_t subsititions, which is why it's acting as a guinea pig.

Any bugs may not be immediately apparent as they are generally in -va
output, but the testsuite is still fine.