* configure.ac: don't generate sysgnash{plugin,}rc, let           'make' take care of that.         * libbase/Makefile.am: install sysgnash{plugin,}rc, generate           on make, clean on make clean.
From now on, the NPAPI plugin will by default start in stopped mode
(as per syswide config). As long as the NPAPI calls gtk-gnash (the default)
this should be safe.
The KPARTS plugin won't use sysgnashpluginrc yet, and the KDE gui doesn't
support the paused widget either, so should be ok.