Rename image_base to ImageBase.
Pass pointers to ImageBase class around as much as possible to make
handling different kinds of images (specifically RGB and RGBA) possible.

bitmap_character_def can take either rgb or rgba images.

BitmapMovieDefinition takes an ImageBase and decides which
bitmap_character_def to create.

The DefineBitsJpeg2 tag loader always casts to RGB, as this is the
only kind of image it can handle (checked with an assertion).