Start frame one byte further on for VP6F codec.
I've found no documentation for this, but it is evidently necessary. The
FLVParser code does a lot of seeking and re-reading. 

VP6A and SCREEN codecs still don't work. SCREEN because of mismatch in
expected and actual size, which probably comes from the fact that
FLVParser guesses what the size is and doesn't read it for this format.