Define a sound_handler::InputStream abstract class to use for plugging input channels to a sound_handler (to be renamed sound::Mixer). No implementations yet, but the plan is to turn both "aux" streamers and "active_sound" into subclasses of this.
NOTE: the interface and documenation was reviewed by zou#gnash and
kshishkov#ffmpeg, it took time and wasn't easy. The requirement for 
samples format was based on the current SDL expectances, which match
the assumptions found all over the place in gnash.
Further reviews are very welcome as I proceed to define subclasses.