Always initialize data member on readSOL (unless exception thrown, which is an untested case). Make data a getter/setter. Passes another test.
Expect failures due to a SharedObject not being null when it should be. This
wasn't what the test claimed it was doing, and isn't a regression.

Flush on exit. This *must* happen, so if it does cause memory errors as 
a comment suggest it used to, it has to be fixed in a different way from
just dropping it.

SharedObject.flush() can take an argument, so log unimplemented.