Move NamingPolicy to its own files in libbase.
StreamProvider is no longer a singleton and must be passed to the RunInfo.
The NamingPolicy for cache files may be passed to the StreamProvider (either
in the ctor or using setNamingPolicy()).

RunInfo pointers are now shared_ptrs to signify shared ownership; this
reduces the importance of construction/destruction order and should
make future refactoring easier.

LoadVariablesThread takes a StreamProvider in the ctor in order to fetch
its stream.