Begin rearranging inheritance to reflect AS3 classes better; this makes handling StaticText objects easier for TextSnapshot, and also prepares for implementing AS3.
generic_character becomes DisplayObject, from which Shape and StaticText
inherit directly.

character should derive from DisplayObject as InteractiveDisplayObject,
not be its base class, but this is not yet implemented.

All character_defs now require a createDisplayObject() method. This
unfortunately includes BitmapMovieDefinition and SWFMovieDefinition, but
that reflects the existing bad design.