Don't use a separate naming style for AVM2 names in VM.
Implement SymbolClassTag fully (but not tested properly).

Check file attributes for whether to use AVM2, and notify when the wrong
tags are present. Currently we also throw a parsing exception to avoid
adding code of the wrong type, but whether the file attribute is definitive
needs more testing. This attribute should also be used for other VM-specific
code (e.g. TextField constructor) in place of SWF version, which is wrong.