Fix most AS interface errors.
Microphone has a constructor that returns an object with some microphone
properties.

Microphone::get() is not a constructor, but a static accessor, so should
be attached to the class (constructor function).

Non-static AS members should not be attached to the class
(constructor function).

Properties are added to the prototype when Microphone::get is called.

Make the class function non-static.