Make some attempt to write a proper interface for VideoInput.
This involves significant changes to VideoInputGst, including the tests. The
test now only uses the public interface of VideoInput (except in a few
cases).

Implementation-specific requirements are not tested. Instead assertions are
used in the implementation.

Tests for bad values have been removed. There is no way to set dimensions from
ActionScript. Instead, a width/height is requested and the VideoInput class
should select an appropriate available one.

The tests should be replaced with generic tests when this functionality is
implemented.