Implement AudioInput using a proper interface, without ifdefs, and without all the awful conversions in user code.
Add AudioInputFfmpeg as a faked class (not that Gst is any better, just
more of a mess).

Expect failure in Microphone.as; it's not worth the ugly hack to get that
pass.

Disable libmedia.all tests for audio and video input; they take a long time,
have no conceivable use for implementing Camera or Microphone classes, and
merely test that a poor design stays as it is. Both Audio and Video input
subclasses need to be redesigned from scratch after actionscript requirements
are clear, not before!