Separate AVM2 more completely from AVM1, including:
- Do not use Namespaces in AVM1 ClassHierarchy. Indeed there is little point
  in using ClassHierarchy for AVM1.
- Rename Class to Script, as it is a static description of a Class, but not
  itself a Class.
- Move Namespace into the abc subdir and do not build it when AVM2 is disabled.
- Do not use AVM2 resources at all in ClassHierarchy when AVM2 is disabled.

Provide better documentation for AVM2 classes, matching terms to the official
documentation and the terms used in tests.

Namespaces now collect Scripts not Classes. It is not clear whether this is
semantically correct or not, though it is likely that builtin AS3 classes
should be static Scripts rather than script Instances (Classes); the Class can
be created when required.