Fix an glaring and ugly mistake in std::less<ObjectURI>().
Extend ObjectURI use to the super class of builtin scripts. Beforehand it was
initialized without a namespace name, and in order to find it the lookup
of all names was too lax, allowing any match to be returned.

This is fine as long as no class has the same name as a builtin, but will
be an almighty mess as soon as that happens.

Be strict about name lookup: the correct namespace is always required.
To do this without too much noise we use a single list of known classes
in name-namespace pairs to look up the required class.

Add a default argument to ObjectURI constructor so that name-only
lookup uses the global namespace. This makes AS2 code more concise. It
also provides a superior replacement to the default argument to most
property lookup functions (still to do).

Use ObjectURI itself as the multi-index container key to reduce code
verbosity when dealing with name/namespace pairs.

Add all used AS3 classes to namedStrings to avoid using string_table (using
it would only require a small change in Globals.cpp).