Document various classes and functions that are central to how the VM runs, but had no explanation whatsoever.
Simplify various functions, currently at the cost of providing action logging
(but this can be restored later).
      
Move global registers to the VM instead of the horrible as_enviroment class,
which additionally means they aren't recreated for every new as_environment.

Introduce the UserFunction as a base for swf_function and builtin_function;
these are identical in the reference implementation.