Use a variable to switch between 'catch-up-time' mode and 'smooth-animation' mode. There are really two different use cases for those two modes:  - A/V sync will benefit from first  - gaming (interaction) will benefit from second due to smoothness
Not sure we can find a compromise here, but PP is reported to take the latter
choice (aka 'elastic time'). 

The variable is to make it easier for developers to see the effect
against their movies and eventually expose it as a compile-time or
build-time configuration.

The default is the current "catch-up-time" (for A/V sync) but maybe
we should change or think about a mixed approach