npapi/writelauncher: Prevent script parameter strings from escaping shell quoting
Make sure a ' character in the parameter string doesn't escape the shell
quoting. This is accomplished by replacing it with a '\'' sequence, i.e.
ending the single quoting, including an explicitely quoted single quote
character, and starting the quoting again.

As the substitution using boost's string algorithm library is trivial,
just do it inline it the script writing statements.