Configure generates a script that installs everything it suggests.
All the ".deb users install this" lines have been replaced with macros that automatically
format suggestions on how to install software, whether you are using .deb or .rpm or just
plain source code. After running configure, the user will be able to run a simple command
as root, and the build system will attempt to install anything that configure wants.
This version should work perfectly on .deb platforms and should mostly work on .rpm
platforms. This version should install everything needed in one try, as it also fixes a
few bugs in configure.ac. Since all information about packages are now encoded in macros,
extending and modifying the system is as simple as changing the macros declaration.
This patch has been successfully tested on Mint 11 (natty) and tested a bit on Fedora 15.
There may be bugs in the handling of some packages that were weirdly checked (such as
FFmpeg, boost, and the like. The checking wasn't consistent and I may have caused bugs).