Fix button key listeners.
Only one listener can be registered for each key. Subsequent attempts
are ignored. A frame count is needed for skip back, because new
DisplayObjects are constructed before the old ones are destroyed when
jumping back, and this means that their attempt to register a key
listener is blocked by the not-yet-destroyed Button.