Pass SoundInfos by value.
They aren't very big, don't get passed very often, and this saves a load
of code required for the auto_ptrs.

Force implementation of sound_handler::create_sound in base classes by
making it pure virtual with an implementation. This should help
externally-maintained parts to adapt to the interface changes.