Don't memcpy data if you need to endian-swap it anyways.
Formerly, the code memcpy'd the data to a new buffer,
and then endian-swapped it, thus doing twice the work.
Now it should do the endian-swapping during the copy phase.
The compiler will usually optimize the byte swap into a single
rot8 instruction - I assume BE CPUs have the same instruction.