Make ConstantPool managed by VM
These changes decouple ConstantPool objects from action_buffer
allowing them to be used across multiple DoAction blocks in a
single frame. Also associates a ConstantPool to each SWF-defined
function so that its execution is not dependent on caller
environment.

All changes are tested with focused tests in misc-swfmill.all,
which now pass except for a single case (dict_event.swf).
An swfdec testcase also succeeds.

The still-failing test seems to suggest that clip event code should
NOT be run in a pool sandbox but doing so breaks other swfdec tests.