Allow abandoning synchonization.
If A/V synchronization problems are due to ActionScript execution rather
than rendering, the chances are we'll never catch up to the sound. In
this case we notify the user after the timeout limit is reached and
check if we should abandon the attempt to synchonize.

Make the movie_root::abortOnScriptTimeout() function into an all-purpose
queryInterface() function so we can reuse it.