Revert "Initialize NetStream prototype only once"
This reverts commit 404343a510a88352b465ae37026102cd1df705e0.

Statics are not allowed, if we want to do this it'll need
a flag in VM or something like that.

Ideally we should have tests checking the behavior more closely,
in particular if it is possible for users to somehow delete those
properties from the prototype we should test what happens on next
NetStream initialization.