Make sure the MediaFactory singleton really is a singleton.
GnashFactory singletons aren't guaranteed to be singletons,
because the compiler creates a fresh template instance whenever
a template is fully specified. This is normally resolved at link
time for (at least for templated function with local statics),
when GCC chucks out all but one instantiation of the template
class.

Unfortunately this doesn't work with Clang, which happily creates
two instances, one for each DSO that instantiates the template, I
suspect, and the ambiguity is not resolved by the linker.

We now explicitly instantiate the template class in
MediaHandler.cpp, which is made possible by C++11's extern
template.

With this change, all tests that rely on a working media handler
now pass with a Clang-built Gnash.