Switch the GcResource container type to forward_list.
We don't make use of bidirectional access, so a singly linked list
is adequate. The order of elements seems to be irrelevant, so
it doesn't matter that we now push elements at the front rather
than the back.

This change reduces the needed memory for this container (that
typically contains a few thousand pointers) by about two-thirds.
cleanUnreachable() runs in slightly less time.