Reduce malloc() churn for SWF-embedded audio.
Decoded data from audio sources stored inside SWFs are decoded and
stored in a growing buffer. This buffer is appended to as small
chunks of data become available. In order to store the additional
data, a new buffer big enough to contain all the data is allocated.
A minute of audio corresponds to about 10MB of decoded data. The
ensuing (re)allocation of such a large buffer adds significantly to
the resource pressure on the system.

This change alters the situation so that instead of allocating a new
buffer, each buffer is added to a list as-is, at the cost of having
to keep track of reads in more detail.