Savannah #43198. Work around boost 1.56+ incompatibility.
A breaking change was introduced in boost 1.56 [1]: the implicit
conversion to bool was removed. Instead, an explicit operator
bool was provided. The getters were also changed, so we cannot
use them and retain compatibility with older boost versions.

By putting the would-be boolean expression in a conditional, we
now invoke the explicit operator bool (also known as contextual
conversion to bool) for boost >= 1.56 and the implicit conversion
to bool for < 1.56.

[1] https://svn.boost.org/trac/boost/ticket/4227